ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.0",
    "fastmcp>=2.13.0.2",
    "joblib>=1.5.2",
    "langchain>=1.0.3",
    "langchain-core>=1.0.3",
    "langchain-groq>=1.0.0",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-text-splitters>=1.0.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "uvicorn[standard]>=0.38.0",
]
